+++ zcash_client_sqlite/src/testing/pool.rs
@@ -57,6 +57,13 @@ pub(crate) fn spend_everything_single_step_proposed_transfer<T: ShieldedPoolTest
     )
 }
 
+#[cfg(not(feature = "transparent-inputs"))]
+pub(crate) fn send_max_to_tex_fails_without_transparent_inputs<T: ShieldedPoolTester>() {
+    zcash_client_backend::data_api::testing::pool::send_max_to_tex_fails_without_transparent_inputs::<
+        T,
+    >(TestDbFactory::default(), BlockCache::new())
+}
+
 #[cfg(feature = "transparent-inputs")]
 pub(crate) fn send_max_spendable_to_transparent<T: ShieldedPoolTester>() {
     zcash_client_backend::data_api::testing::pool::send_max_spendable_to_transparent::<T>(